rename blockchain package to blocksync
