+++ proto/tendermint/blocksync/message.go
@@ -1,4 +1,4 @@
-package blockchain
+package blocksync
 
 import (
 	"errors"
