+++ proto/tendermint/blocksync/types.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package tendermint.blockchain;
 
-option go_package = "github.com/tendermint/tendermint/proto/tendermint/blockchain";
+option go_package = "github.com/tendermint/tendermint/proto/tendermint/blocksync";
 
 import "tendermint/types/block.proto";
